Create common components folder
