--- src/features/common/Form/index.ts
@@ -1,5 +1,0 @@
-export { Input } from "./Input";
-export { Error } from "./Error";
-export { Textarea } from "./Textarea";
-export { Label } from "./Label";
-export { FormInput } from "./FormInput";
