+++ src/features/common/components/BackButton.tsx
@@ -1,0 +1,12 @@
+import { ChevronLeftCircle } from "lucide-react";
+import { useRouter } from "next/router";
+
+export const BackButton = () => {
+  const router = useRouter();
+
+  return (
+    <button onClick={() => router.back()} className="mr-auto">
+      <ChevronLeftCircle size={32} />
+    </button>
+  );
+};
