+++ src/features/common/components/Form/index.ts
@@ -1,0 +1,5 @@
+export { Input } from "./Input";
+export { Error } from "./Error";
+export { Textarea } from "./Textarea";
+export { Label } from "./Label";
+export { FormInput } from "./FormInput";
